+++ pydfs_lineup_optimizer/sites/fanball/__init__.py
@@ -1,5 +1,5 @@
-from pydfs_lineup_optimizer.sites.fanball.settings import FanBallNRLSettings
+from pydfs_lineup_optimizer.sites.fanball.settings import FanBallFOOTBALLSettings
 
 __all__ = [
-    'FanBallNRLSettings',
+    'FanBallFOOTBALLSettings',
 ]